Renaming files to snake case
