+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:omni_manager/pages/Home.dart';
-import 'package:omni_manager/pages/Login.dart';
-import 'package:omni_manager/pages/Register.dart';
+import 'package:omni_manager/pages/home.dart';
+import 'package:omni_manager/pages/login.dart';
+import 'package:omni_manager/pages/register.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'utils/constants.dart';
